requirements.txt
pytest==6.2.1
requests==2.24.0
beautifulsoup4==4.9.3
python-dateutil==2.8.1